some file refine and move style into their files
