+++ app/web/src/@types/react-app-env.d.ts
@@ -1,31 +1,1 @@
 /// <reference types="react-scripts" />
-
-// declare module 'redux-persist/integration/react'
-declare module 'redux-persist/integration/react' {
-  import { ReactNode, PureComponent } from 'react'
-  import { Persistor } from 'redux-persist/es/types'
-
-  /** @see PersistGate */
-  interface PersistGateProps {
-    persistor: Persistor
-    onBeforeLift?(): void | Promise<void>
-    children?: ReactNode | ((bootstrapped: boolean) => ReactNode)
-    loading?: ReactNode
-  }
-
-  /** @see PersistGate */
-  interface PersistorGateState {
-    bootstrapped: boolean
-  }
-
-  /**
-   * Entry point of your react application to allow it persist a given store @see Persistor and its configuration.
-   * @see Persistor
-   * @see PersistGateProps
-   * @see PersistGateState
-   */
-  class PersistGate extends PureComponent<
-    PersistGateProps,
-    PersistorGateState
-  > {}
-}
