+++ app/web/src/components/common/Header/Header.component.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
+
 import { history } from '../../../util/'
 import { Container, NavRow, Logo, NavItems, Item } from './Header.styled'
