--- app/web/src/components/layout/UploadDownload/worker.js
@@ -1,4 +1,0 @@
-window.onmessage = function (e) {
-  // console.log(e)
-  this.postMessage('hello from worker')
-}
