+++ app/web/src/util/randomKey.util.ts
@@ -6,5 +6,6 @@ export function makeKey() {
   for (var i = 0; i < 20; i++)
     text += possible.charAt(Math.floor(Math.random() * possible.length))
 
+  // encrypt base65-1
   return btoa(text)
 }
